Data of each Item is now passed to the ItemPage when clicked and swaps view.
Item Page doesn't show up properly but data successfully enters textViews.

+++ app/src/main/java/ie/ul/cs4084project/Item.java
@@ -10,13 +10,13 @@ public class Item {
 
     private String name;
     private String description;
-    private int price;
+    private double price;
     @ServerTimestamp
     private Date timeStamp;
 
     public Item() { }
 
-    public Item(String name, String description, int price) {
+    public Item(String name, String description, double price) {
         this.name = name;
         this.description = description;
         this.price = price;
@@ -24,11 +24,11 @@ public class Item {
 
     public String getName() { return name; }
     public String getDescription() { return description; }
-    public int getPrice() { return price; }
+    public double getPrice() { return price; }
     public Date getTimeStamp() { return timeStamp; }
 
     public void setName(String name) { this.name = name; }
     public void setDescription(String description) { this.description = description; }
-    public void setPrice(int price){ this.price = price; }
+    public void setPrice(double price){ this.price = price; }
     public void setTimeStamp(Date timeStamp) { this.timeStamp = timeStamp; }
 }
